gamepad test rework
